requirements.txt
flask>=3.0.0
fastai>=2.7.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
